refactor(beacon-node): simplify libp2p init
