+++ packages/beacon-node/src/network/index.ts
@@ -1,7 +1,7 @@
 export * from "./events.js";
 export * from "./interface.js";
 export * from "./network.js";
-export * from "./nodejs/index.js";
+export * from "./libp2p/index.js";
 export * from "./gossip/index.js";
 export * from "./reqresp/ReqRespBeaconNode.js";
 export * from "./util.js";
