+++ packages/beacon-node/src/network/libp2p/noise.ts
