--- packages/beacon-node/src/network/nodejs/index.ts
@@ -1,2 +1,0 @@
-export * from "./bundle.js";
-export * from "./util.js";
